+++ src/backend/marsha/core/api/video.py
@@ -58,6 +58,7 @@ from ..utils.medialive_utils import (
     update_id3_tags,
     wait_medialive_channel_is_created,
 )
+from ..utils.send_emails import send_ready_to_convert_notification
 from ..utils.time_utils import to_timestamp
 from ..utils.xmpp_utils import close_room, create_room, reopen_room_for_vod
 from .base import APIViewMixin, BulkDestroyModelMixin, ObjectPkMixin
@@ -751,6 +752,7 @@ class VideoViewSet(
             video.live_state = defaults.STOPPED
             live_info.update({"stopped_at": stamp})
             video.live_info = live_info
+            send_ready_to_convert_notification(video)
 
         if serializer.validated_data["state"] == defaults.HARVESTED:
             video.live_state = defaults.HARVESTED